Extracted common parts of lesson test cases
